@@ -1,18 +1,15 @@
 {
     "config": {
         "platform": {
-            "php": "5.6.29"
+            "php": "7.0.8"
         }
     },
     "require-dev": {
-        "phpunit/phpunit": "^5.6",
         "bamarni/composer-bin-plugin": "^1.2",
-        "jakub-onderka/php-parallel-lint": "^1.0",
-        "zendframework/zend-ldap": "^2.8",
-        "owncloud/coding-standard": "^1.0"
+        "zendframework/zend-ldap": "^2.8"
     },
     "require": {
-        "php": ">=5.6",
+        "php": ">=7.0.8",
         "ext-ldap": "*"
     },
     "extra": {
