@@ -1,7 +1,7 @@
 {
     "config" : {
         "platform": {
-            "php": "5.6.33"
+            "php": "7.1"
         }
     },
     "require": {
@@ -12,6 +12,9 @@
         "jarnaiz/behat-junit-formatter": "^1.3",
         "rdx/behat-variables": "^1.2",
         "sensiolabs/behat-page-object-extension": "^2.0",
-        "symfony/translation": "^3.4"
+        "symfony/translation": "^3.4",
+        "sabre/xml": "^2.1",
+        "guzzlehttp/guzzle": "^5.3",
+        "phpunit/phpunit": "^7.5"
     }
 }
