@@ -39,15 +39,15 @@ use OCA\User_LDAP\User\Manager;
  */
 class Group_LDAPTest extends \Test\TestCase {
 	/**
-	 * @var Access|\PHPUnit_Framework_MockObject_MockObject
+	 * @var Access|\PHPUnit\Framework\MockObject\MockObject
 	 */
 	protected $access;
 	/**
-	 * @var Connection|\PHPUnit_Framework_MockObject_MockObject
+	 * @var Connection|\PHPUnit\Framework\MockObject\MockObject
 	 */
 	protected $connection;
 
-	public function setUp() {
+	public function setUp(): void {
 		parent::setUp();
 		$this->connection = $this->createMock(Connection::class);
 
