@@ -1,35 +1,166 @@
 OC.L10N.register(
     "user_ldap",
     {
-    "Failed to clear the mappings." : "فشل مسح الارتباطات (mappings)",
-    "Failed to delete the server configuration" : "تعذر حذف ملف إعدادات الخادم",
-    "The configuration is valid and the connection could be established!" : "الإعدادت صحيحة",
-    "The configuration is valid, but the Bind failed. Please check the server settings and credentials." : "الإعدادات صحيحة، لكن لم ينجح الارتباط. يرجى التأكد من إعدادات الخادم وبيانات التحقق من الدخول.",
-    "The configuration is invalid. Please have a look at the logs for further details." : "الإعدادات غير صحيحة. يرجى الاطلاع على سجلات المتابعة للمزيد من التفاصيل.",
-    "No action specified" : "لم يتم تحديد الإجراء",
-    "No configuration specified" : "لم يتم تحديد الإعدادات.",
-    "No data specified" : "لم يتم تحديد البيانات.",
-    " Could not set configuration %s" : "تعذر تنفيذ الإعداد %s",
-    "Configuration incorrect" : "الإعدادات غير صحيحة",
-    "Configuration incomplete" : "الإعدادات غير مكتملة",
-    "Configuration OK" : "الإعدادات صحيحة",
-    "Select groups" : "إختر مجموعة",
-    "Select object classes" : "اختر أصناف المكونات",
-    "{nthServer}. Server" : "الخادم {nthServer}.",
-    "Do you really want to delete the current Server Configuration?" : "هل ترغب فعلاً في حذف إعدادات الخادم الحالي؟",
+    "The Base DN appears to be wrong" : "يبدو أن الاسم المميز الأساسي خاطئ",
+    "Testing configuration…" : "جارٍ اختبار التكوين...",
+    "Configuration incorrect" : "التكوين غير صحيح",
+    "Configuration incomplete" : "التكوين غير مكتمل",
+    "Configuration OK" : "التكوين صحيح",
+    "Select groups" : "تحديد مجموعات",
+    "Select object classes" : "تحديد فئات العنصر",
+    "Please check the credentials, they seem to be wrong." : "يُرجى فحص بيانات الاعتماد، يبدو أنها خاطئة.",
+    "Please specify the port, it could not be auto-detected." : "يُرجى تحديد المنفذ، يتعذّر الكشف عنه تلقائيًا.",
+    "Base DN could not be auto-detected, please revise credentials, host and port." : "يتعذّر الكشف عن الاسم المميز الأساسي تلقائيًا، يُرجى مراجعة بيانات الاعتماد، والمضيف، والمنفذ.",
+    "Could not detect Base DN, please enter it manually." : "تعذّر الكشف عن الاسم المميز الأساسي، يُرجى إدخال معلوماته يدويًا.",
+    "{nthServer}. Server" : "خادم {nthServer}.",
+    "No object found in the given Base DN. Please revise." : "لم يتم العثور على أي عنصر في الاسم المميز الأساسي المقدم. يُرجى مراجعة معلوماته.",
+    "More than 1,000 directory entries available." : "يتوفر أكثر من 1000 إدخال للدليل.",
+    " entries available within the provided Base DN" : "تتوفر الإدخالات في الاسم المميز الأساسي المُقدم",
+    "An error occurred. Please check the Base DN, as well as connection settings and credentials." : "حدث خطأ ما. يُرجى فحص الاسم المميز الأساسي، وإعدادات الاتصال، وبيانات الاعتماد.",
+    "Do you really want to delete the current Server Configuration?" : "هل ترغب حقًا في حذف تكوين الخادم الحالي؟",
     "Confirm Deletion" : "تأكيد الحذف",
-    "Select attributes" : "اختر الخصائص",
-    "_%s group found_::_%s groups found_" : ["لا توجد مجموعات: %s","تم إيجاد %s مجموعة واحدة","تم إيجاد %s مجموعتين","تم إيجاد %s مجموعات","تم إيجاد %s مجموعة","تم إيجاد %s مجموعة/مجموعات"],
+    "Mappings cleared successfully!" : "نجح مسح التعيينات!",
+    "Error while clearing the mappings." : "حدث خطأ أثناء مسح التعيينات.",
+    "Anonymous bind is not allowed. Please provide a User DN and Password." : "غير مسموح بالربط مجهول المصدر. يُرجى توفير الاسم المميز للمستخدم وكلمة المرور.",
+    "LDAP Operations error. Anonymous bind might not be allowed." : "خطأ في عمليات LDAP. غير مسموح بالربط مجهول المصدر.",
+    "Saving failed. Please make sure the database is in Operation. Reload before continuing." : "فشل الحفظ. يُرجى التأكد من أن قاعدة البيانات قيد التشغيل. يُرجى إعادة التحميل قبل المتابعة.",
+    "Switching the mode will enable automatic LDAP queries. Depending on your LDAP size they may take a while. Do you still want to switch the mode?" : "سيؤدي تبديل الوضع إلى تمكين استفسارات LDAP التلقائية. وستستغرق هذه العملية بعض الوقت حسب حجم LDAP. ألا تزال ترغب في تبديل الوضع؟",
+    "Mode switch" : "تبديل الوضع",
+    "Select attributes" : "تحديد السمات",
+    "User not found. Please check your login attributes and username. Effective filter (to copy-and-paste for command line validation): <br/>" : "لم يتم العثور على المستخدم. يُرجى التحقق من سمات تسجيل الدخول واسم المستخدم الخاص بك. عامل تصفية فعال (للنسخ واللصق للتحقق من صحة سطر الأوامر): <br/>",
+    "User found and settings verified." : "تم العثور على مستخدم والتحقق من صحة الإعدادات.",
+    "Settings verified, but more than one user was found. Only the first will be able to login. Consider a more narrow filter." : "تم التحقق من صحة الإعدادات، لكن تم العثور على أكثر من مستخدم واحد. سيتمكن المستخدم الأول فقط من تسجيل الدخول. استخدم عامل تصفية أدق.",
+    "An unspecified error occurred. Please check the settings and the log." : "حدث خطأ غير محدد. يُرجى التحقق من الإعدادات والسجل.",
+    "The search filter is invalid, probably due to syntax issues like uneven number of opened and closed brackets. Please revise." : "عامل تصفية البحث غير صالح، ربما بسبب مشكلات في بناء الجملة مثل عدم تساوي عدد الأقواس المفتوحة والمغلقة. يُرجى مراجعة ذلك.",
+    "A connection error to LDAP / AD occurred, please check host, port and credentials." : "حدث خطأ في الاتصال بـ LDAP / AD، يُرجى فحص المضيف، والمنفذ، وبيانات الاعتماد.",
+    "The %uid placeholder is missing. It will be replaced with the login name when querying LDAP / AD." : "العنصر النائب %uid مفقود. سيتم استخدام اسم تسجيل الدخول بدلًا منه عند الاستعلام عن LDAP / AD.",
+    "Please provide a login name to test against" : "يُرجى تقديم اسم تسجيل الدخول لاختبار",
+    "The group box was disabled, because the LDAP / AD server does not support memberOf." : "تم تعطيل خانة المجموعة؛ لأن خادم LDAP / AD لا يدعم memberOf.",
     "Server" : "خادم",
-    "Users" : "المستخدمين",
+    "Users" : "مستخدمون",
+    "Login Attributes" : "سمات تسجيل الدخول",
     "Groups" : "مجموعات",
+    "The configuration is invalid: anonymous bind is not allowed." : "التكوين غير صالح: غير مسموح بالربط مجهول المصدر.",
+    "The configuration is valid and the connection could be established!" : "التكوين صالح ويمكن إنشاء اتصال!",
+    "The configuration is valid, but the Bind failed. Please check the server settings and credentials." : "التكوين صالح، لكن فشل الربط. يُرجى التحقق من إعدادات الخادم وبيانات اعتماده.",
+    "The configuration is invalid. Please have a look at the logs for further details." : "التكوين غير صالح. يُرجى الاطلاع على السجلات للمزيد من التفاصيل.",
+    "Failed to delete the server configuration" : "فشل حذف تكوين الخادم",
+    "Failed to clear the mappings." : "فشل مسح التعيينات.",
+    "No data specified" : "لم يتم تحديد أي بيانات",
+    " Could not set configuration %s" : "لا يمكن تعيين التكوين %s",
+    "Action does not exist" : "الإجراء غير موجود",
+    "_%s group found_::_%s groups found_" : ["تم العثور على %s مجموعة","تم العثور على %s مجموعات","تم العثور على %s مجموعات","تم العثور على %s مجموعات","تم العثور على %s مجموعات","تم العثور على %s مجموعات"],
+    "_%s user found_::_%s users found_" : ["تم العثور على %s مستخدم","تم العثور على %s مستخدمين","تم العثور على %s مستخدمين","تم العثور على %s مستخدمين","تم العثور على %s مستخدمين","تم العثور على %s مستخدمين"],
+    "Could not detect user display name attribute. Please specify it yourself in advanced ldap settings." : "لا يمكن الكشف عن سمة اسم عرض المستخدم. يُرجى تحديد ذلك بنفسك في إعدادات ldap المتقدمة.",
+    "Could not find the desired feature" : "تعذّر العثور على الميزة المطلوبة",
+    "Invalid Host" : "مضيف غير صالح",
+    "Test Configuration" : "اختبار التكوين",
     "Help" : "المساعدة",
+    "Groups meeting these criteria are available in %s:" : "تتوفر المجموعات التي تستوفي هذه المعايير في %s:",
+    "Only these object classes:" : "فئات العناصر هذه فقط:",
+    "Only from these groups:" : "من هذه المجموعات فقط:",
+    "Search groups" : "البحث عن المجموعات",
+    "Available groups" : "المجموعات المتوفرة",
+    "Selected groups" : "المجموعات المحددة",
+    "Edit LDAP Query" : "تحرير استعلام LDAP",
+    "LDAP Filter:" : "عامل تصفية LDAP:",
+    "The filter specifies which LDAP groups shall have access to the %s instance." : "يحدد عامل التصفية مجموعات LDAP التي يجب أن تدخل إلى مثيل %s.",
+    "Verify settings and count groups" : "التحقق من صحة الإعدادات وحساب المجموعات",
+    "When logging in, %s will find the user based on the following attributes:" : "عند تسجيل الدخول، %s سيعثر على المستخدم حسب السمات التالية:",
+    "LDAP / AD Username:" : "اسم مستخدم LDAP / AD:",
+    "Allows login against the LDAP / AD username, which is either uid or samaccountname and will be detected." : "يسمح بتسجيل الدخول باسم المستخدم LDAP / AD، الذي سيكون إمّا uid وإمّا samaccountname، وسيتم الكشف عنه.",
+    "LDAP / AD Email Address:" : "عنوان البريد الإلكتروني لـ LDAP / AD:",
+    "Allows login against an email attribute. Mail and mailPrimaryAddress will be allowed." : "يسمح بتسجيل الدخول بسمة البريد الإلكتروني. سيتم السماح بعنوان البريد وmailPrimaryAddress.",
+    "Other Attributes:" : "سمات أخرى:",
+    "Defines the filter to apply, when login is attempted. %%uid replaces the username in the login action. Example: \"uid=%%uid\"" : "يحدد عامل التصفية الذي يتم استخدامه، عند محاولة تسجيل الدخول. %%uid يستبدل اسم المستخدم في إجراء تسجيل الدخول. على سبيل المثال: \"uid=%%uid\"",
+    "Test Loginname" : "اختبار اسم تسجيل الدخول",
+    "Verify settings" : "التحقق من صحة الإعدادات",
+    "1. Server" : "1. الخادم",
+    "%s. Server:" : "%s. الخادم:",
+    "Add a new and blank configuration" : "إضافة تكوين جديد وفارغ",
+    "Copy current configuration into new directory binding" : "نسخ التكوين الحالي في ربط دليل جديد",
+    "Delete the current configuration" : "حذف التكوين الحالي",
     "Host" : "المضيف",
-    "Port" : "المنفذ",
+    "You can omit the protocol, except you require SSL. Then start with ldaps://" : "يمكنك حذف البروتوكول، باستثناء حالة طلبك SSL. ثم ابدأ بـ ldaps://",
+    "Port" : "منفذ",
+    "Detect Port" : "اكتشاف منفذ",
+    "User DN" : "الاسم المميز للمستخدم",
+    "The DN of the client user with which the bind shall be done, e.g. uid=agent,dc=example,dc=com. For anonymous access, leave DN and Password empty." : "اسم مستخدم العميل الذي يتم به الربط، على سبيل المثال، uid=agent,dc=example,dc=com. للدخول مجهول المصدر، اترك خانتَي الاسم المميز وكلمة المرور فارغتَين.",
     "Password" : "كلمة المرور",
-    "Back" : "رجوع",
-    "Continue" : "المتابعة",
-    "Advanced" : "تعديلات متقدمه",
-    "Email Field" : "خانة البريد الإلكتروني"
+    "For anonymous access, leave DN and Password empty." : "للدخول مجهول المصدر، اترك خانتَي الاسم المميز وكلمة المرور فارغتَين.",
+    "One Base DN per line" : "اسم مميز أساسي واحد لكل سطر",
+    "You can specify Base DN for users and groups in the Advanced tab" : "يمكنك تحديد الاسم المميز الأساسي للمستخدمين والمجموعات في علامة التبويب المتقدمة",
+    "Detect Base DN" : "اكتشاف الاسم المميز الأساسي",
+    "Test Base DN" : "اختبار الاسم المميز الأساسي",
+    "Avoids automatic LDAP requests. Better for bigger setups, but requires some LDAP knowledge." : "يتجنب طلبات LDAP التلقائية. وهو أفضل للإعدادات الأكبر، لكن يتطلب بعض المعرفة بـ LDAP.",
+    "Manually enter LDAP filters (recommended for large directories)" : "إدخال عوامل تصفية LDAP يدويًا (موصى به للأدلة الكبيرة)",
+    "%s access is limited to users meeting these criteria:" : "وصول %s مقتصر على المستخدمين الذين يتوافقون مع هذه المعايير:",
+    "The most common object classes for users are organizationalPerson, person, user, and inetOrgPerson. If you are not sure which object class to select, please consult your directory admin." : "فئات العناصر الأكثر شيوعًا للمستخدمين هي organisationalPerson، والشخص، والمستخدم، وinetOrgPerson. وإذا لم تكن متأكدًا من فئة الكائن المراد تحديدها، فيُرجى الرجوع إلى مسؤول الدليل.",
+    "The filter specifies which LDAP users shall have access to the %s instance." : "يحدد عامل التصفية مستخدمي LDAP الذين يجب أن يدخلوا إلى مثيل %s.",
+    "Verify settings and count users" : "التحقق من صحة الإعدادات وحساب المستخدمين",
+    "Saving" : "حفظ",
+    "Back" : "عودة",
+    "Continue" : "متابعة",
+    "LDAP" : "LDAP",
+    "Expert" : "خبير",
+    "Advanced" : "الإعدادات المتقدمة",
+    "<b>Warning:</b> Apps user_ldap and user_webdavauth are incompatible. You may experience unexpected behavior. Please ask your system administrator to disable one of them." : "<b>تحذير:</b> التطبيقان user_ldap وuser_webdavauth غير متوافقَين. قد تتعرض إلى طريقة غير متوقعة. يُرجى مطالبة مسؤول النظام بتعطيل أحدهما.",
+    "<b>Warning:</b> The PHP LDAP module is not installed, the backend will not work. Please ask your system administrator to install it." : "<b>تحذير:</b> لم يتم تثبيت وحدة PHP LDAP، ولن تعمل الواجهة الخلفية. يُرجى مطالبة مسؤول النظام بتثبيتها.",
+    "Connection Settings" : "إعدادات الاتصال",
+    "Configuration Active" : "التكوين نشط",
+    "When unchecked, this configuration will be skipped." : "في حالة عدم التحديد، سيتم تجاوز هذا التكوين.",
+    "Backup (Replica) Host" : "مضيف النسخ الاحتياطي (نسخة مماثلة)",
+    "Give an optional backup host. It must be a replica of the main LDAP/AD server." : "قدِّم مضيف نسخ احتياطيًّا اختياريًا. يجب أن يكون نسخة مماثلة من خادم LDAP/AD الرئيسي.",
+    "Backup (Replica) Port" : "منفذ النسخ الاحتياطي (نسخة مماثلة)",
+    "Disable Main Server" : "تعطيل الخادم الرئيسي",
+    "Only connect to the replica server." : "يُرجى الاتصال بالخادم المماثل فقط.",
+    "Turn off SSL certificate validation." : "أوقف تشغيل التحقق من صحة شهادة SSL.",
+    "Not recommended, use it for testing only! If connection only works with this option, import the LDAP server's SSL certificate in your %s server." : "لا يوصى بذلك، استخدمها للاختبار فقط! إذا عمل الاتصال مع هذا الخيار فقط، فاسترد شهادة SSL لخادم LDAP في خادم %s.",
+    "Cache Time-To-Live" : "فترة بقاء ذاكرة التخزين المؤقت",
+    "in seconds. A change empties the cache." : "بالثواني. أي تغيير سيفرغ ذاكرة التخزين المؤقت.",
+    "Network Timeout" : "مهلة الشبكة",
+    "timeout for all the ldap network operations, in seconds." : "مهلة جميع عمليات شبكة ldap بالثواني.",
+    "Directory Settings" : "إعدادات الدليل",
+    "User Display Name Field" : "خانة اسم عرض المستخدم",
+    "The LDAP attribute to use to generate the user's display name." : "سمة LDAP المراد استخدامها لإنشاء اسم عرض المستخدم.",
+    "2nd User Display Name Field" : "خانة اسم عرض المستخدم الثاني",
+    "Optional. An LDAP attribute to be added to the display name in brackets. Results in e.g. »John Doe (john.doe@example.org)«." : "اختياري. سمة LDAP المُراد إضافتها إلى اسم العرض بين الأقواس. النتيجة على سبيل المثال، »John Doe (john.doe@example.org)«.",
+    "Base User Tree" : "شجرة المستخدم الرئيسية",
+    "One User Base DN per line" : "اسم مميز أساسي واحد للمستخدم لكل سطر",
+    "User Search Attributes" : "سمات البحث عن المستخدم",
+    "Optional; one attribute per line" : "اختياري؛ سمة واحدة لكل سطر",
+    "Each attribute value is truncated to 191 characters" : "يتم اقتطاع كل قيمة سمة إلى 191 حرفًا",
+    "Group Display Name Field" : "خانة اسم عرض المجموعة",
+    "The LDAP attribute to use to generate the groups's display name." : "سمة LDAP المراد استخدامها لإنشاء اسم عرض المجموعات.",
+    "Base Group Tree" : "شجرة المجموعة الرئيسية",
+    "One Group Base DN per line" : "اسم مميز أساسي واحد للمجموعة لكل سطر",
+    "Group Search Attributes" : "سمات البحث عن المجموعة",
+    "Group-Member association" : "اقتران أعضاء المجموعة",
+    "Dynamic Group Member URL" : "عنوان URL لعضو مجموعة ديناميكي",
+    "The LDAP attribute that on group objects contains an LDAP search URL that determines what objects belong to the group. (An empty setting disables dynamic group membership functionality.)" : "تحتوي سمة LDAP الموجودة على عناصر المجموعة على عنوان URL للبحث عن LDAP الذي يحدد العناصر التي تنتمي إلى المجموعة. (يعطِّل أي إعداد فارغ وظيفة عضوية المجموعة الديناميكية).",
+    "Nested Groups" : "المجموعات المتداخلة",
+    "When switched on, groups that contain groups are supported. (Only works if the group member attribute contains DNs.)" : "عند تشغيلها، تُدعم المجموعات التي تحتوي على مجموعات. (لا يعمل ذلك إلا إذا كانت سمة عضو المجموعة تحتوي على أسماء مميزة فقط).",
+    "Paging chunksize" : "ترحيل حجم المجموعة",
+    "Chunksize used for paged LDAP searches that may return bulky results like user or group enumeration. (Setting it 0 disables paged LDAP searches in those situations.)" : "يمكنك استخدام حجم المجموعة في عمليات البحث LDAP المُرحَّلة والتي قد تؤدي إلى نتائج ضخمة مثل تعداد المستخدم أو المجموعة. (يؤدي ضبطه إلى 0 إلى تعطيل عمليات بحث LDAP المُرحَّلة في هذه الحالات).",
+    "Special Attributes" : "سمات خاصة",
+    "Quota Field" : "خانة الحصة النسبية",
+    "Leave empty for user's default quota. Otherwise, specify an LDAP/AD attribute." : "اترك خانة الحصة النسبية الافتراضية للمستخدم فارغة. وإلا، فحدد سمة LDAP/AD.",
+    "Quota Default" : "الإعدادات الافتراضية للحصة النسبية",
+    "Override default quota for LDAP users who do not have a quota set in the Quota Field." : "تجاوز الحصة النسبية الافتراضية لمستخدمي LDAP الذين ليست لديهم حصة نسبية محددة في خانة الحصة النسبية.",
+    "Email Field" : "خانة البريد الإلكتروني",
+    "Set the user's email from their LDAP attribute. Leave it empty for default behaviour." : "عيِّن البريد الإلكتروني للمستخدم من سمة LDAP. اترك خانة الطريقة الافتراضية فارغة.",
+    "User Home Folder Naming Rule" : "قاعدة تسمية المجلد الرئيسي للمستخدم",
+    "Leave empty for user name (default). Otherwise, specify an LDAP/AD attribute." : "اترك خانة اسم المستخدم (الافتراضية) فارغة، وإلا فحدد سمة LDAP/AD.",
+    "Internal Username" : "اسم مستخدم داخلي",
+    "By default the internal username will be created from the UUID attribute. It makes sure that the username is unique and characters do not need to be converted. The internal username has the restriction that only these characters are allowed: [ a-zA-Z0-9_.@- ].  Other characters are replaced with their ASCII correspondence or simply omitted. On collisions a number will be added/increased. The internal username is used to identify a user internally. It is also the default name for the user home folder. It is also a part of remote URLs, for instance for all *DAV services. With this setting, the default behavior can be overridden. To achieve a similar behavior as before ownCloud 5 enter the user display name attribute in the following field. Leave it empty for default behavior. Changes will have effect only on newly mapped (added) LDAP users." : "سيتم إنشاء اسم المستخدم الداخلي من سمة UUID افتراضيًا. ويؤكد ذلك أن اسم المستخدم فريد وأن الأحرف لا تحتاج إلى تحويل. ويجب أن يتكون اسم المستخدم الداخلي من هذه الأحرف فقط: [a-zA-Z0-9 _. @ -]. ويتم استخدام أحرف ASCII بدلًا من الأحرف القديمة أو يتم حذفها ببساطة. وفي حالة وجود تضارب، ستتم إضافة/زيادة رقم. ويُستخدَم اسم المستخدم الداخلي لتعريف المستخدم داخليًا. وسيكون أيضًا الاسم الافتراضي للمجلد الرئيسي للمستخدم. وهو جزء من عناوين URL البعيدة، على سبيل المثال، لجميع خدمات DAV *. ويمكن تجاوز الطريقة الافتراضية باستخدام هذا الإعداد. وللحصول على طريقة مماثلة لطريقة ownCloud 5، أدخل سمة اسم عرض المستخدم في الخانة التالية، واترك خانة الطريقة الافتراضية فارغة. ولن تؤثر التغييرات إلا في مستخدمي LDAP المعينين (المضافين) حديثًا.",
+    "Internal Username Attribute:" : "سمة اسم المستخدم الداخلي:",
+    "Override UUID detection" : "تجاوز اكتشاف UUID",
+    "By default, the UUID attribute is automatically detected. The UUID attribute is used to doubtlessly identify LDAP users and groups. Also, the internal username will be created based on the UUID, if not specified otherwise above. You can override the setting and pass an attribute of your choice. You must make sure that the attribute of your choice can be fetched for both users and groups and it is unique. Leave it empty for default behavior. Changes will have effect only on newly mapped (added) LDAP users and groups." : "يتم اكتشاف سمة UUID تلقائيًا بشكل افتراضي. ويتم استخدامها لتحديد مستخدمي ومجموعات LDAP بدون شك. وسيتم إنشاء اسم المستخدم الداخلي أيضًا استنادًا إلى UUID، إذا لم يُذكر خلاف ذلك أعلاه. ويمكنك تخطي الإعداد وتجاوز سمة تختارها. ويجب عليك التأكد من أن السمة التي تختارها يمكن جلبها لكلٍّ من المستخدمين والمجموعات وأنها فريدة. اترك خانة الطريقة الافتراضية فارغة. ولن تؤثر التغييرات إلا في مستخدمي LDAP المعينين (المضافين) حديثًا.",
+    "UUID Attribute for Users:" : "سمة UUID للمستخدمين:",
+    "UUID Attribute for Groups:" : "سمة UUID للمجموعات:",
+    "Username-LDAP User Mapping" : "اسم المستخدم - تعيين مستخدم LDAP",
+    "Usernames are used to store and assign (meta) data. In order to precisely identify and recognize users, each LDAP user will have an internal username. This requires a mapping from username to LDAP user. The created username is mapped to the UUID of the LDAP user. Additionally the DN is cached as well to reduce LDAP interaction, but it is not used for identification. If the DN changes, the changes will be found. The internal username is used all over. Clearing the mappings will have leftovers everywhere. Clearing the mappings is not configuration sensitive, it affects all LDAP configurations! Never clear the mappings in a production environment, only in a testing or experimental stage." : "تُستخدَم أسماء المستخدمين لتخزين البيانات (بيانات التعريف) وتعيينها. ولتحديد المستخدمين والتعرف عليهم بدقة، سيكون لكل مستخدم LDAP اسم مستخدم داخلي. ويتطلب ذلك تعيينًا من اسم المستخدم إلى مستخدم LDAP. ويتم تعيين اسم المستخدم الذي تم إنشاؤه إلى UUID لمستخدم LDAP. وبالإضافة إلى ذلك، يتم تخزين الاسم المميز مؤقتًا لتقليل تفاعل LDAP، لكن لا يتم استخدامه لتحديد الهوية. وفي حالة تغيير الاسم المميز، سيتم العثور على التغييرات، كما يتم استخدام اسم المستخدم الداخلي في جميع المواقف. وسيترك مسح التعيينات آثارًا في كل مكان. فهو ليس حساسًا للتكوين، بل يؤثر في جميع تكوينات LDAP! فلا تمسح التعيينات في بيئة الإنتاج، بل في مرحلة الاختبار أو التجربة فقط.",
+    "Clear Username-LDAP User Mapping" : "مسح اسم المستخدم - تعيين مستخدم LDAP",
+    "Clear Groupname-LDAP Group Mapping" : "مسح اسم المجموعة - تعيين مجموعة LDAP"
 },
 "nplurals=6; plural=n==0 ? 0 : n==1 ? 1 : n==2 ? 2 : n%100>=3 && n%100<=10 ? 3 : n%100>=11 && n%100<=99 ? 4 : 5;");
