@@ -1,68 +1,68 @@
 OC.L10N.register(
     "user_ldap",
     {
-    "The Base DN appears to be wrong" : "Το Base DN φαίνεται να είναι εσφαλμένο",
+    "The Base DN appears to be wrong" : "Το βασικό DN φαίνεται να είναι εσφαλμένο",
     "Testing configuration…" : "Γίνεται δοκιμή ρυθμίσεων...",
-    "Configuration incorrect" : "Η διαμόρφωση είναι λανθασμένη",
-    "Configuration incomplete" : "Η διαμόρφωση είναι ελλιπής",
-    "Configuration OK" : "Η διαμόρφωση είναι εντάξει",
+    "Configuration incorrect" : "Εσφαλμένες ρυθμίσεις",
+    "Configuration incomplete" : "Ελλιπείς ρυθμίσεις",
+    "Configuration OK" : "Οι ρυθμίσεις είναι εντάξει",
     "Select groups" : "Επιλέξτε ομάδες",
     "Select object classes" : "Επιλογή κλάσης αντικειμένων",
-    "Please check the credentials, they seem to be wrong." : "Παρακαλώ ελέγξτε τα στοιχεία, φαίνεται να είναι λάθος.",
-    "Please specify the port, it could not be auto-detected." : "Παρακαλώ ορίστε την θύρα, δεν ήταν εφικτό να εντοπιστεί αυτόματα.",
-    "Base DN could not be auto-detected, please revise credentials, host and port." : "Αδυναμία ανίχνευσης Base DN, παρακαλώ αναθεωρήστε τα διαπιστευτήρια, το διακομιστή και τη θύρα.",
-    "Could not detect Base DN, please enter it manually." : "Αδυναμία ανίχνευσης Base DN, παρακαλώ να το εισάγετε χειροκίνητα.",
+    "Please check the credentials, they seem to be wrong." : "Παρακαλούμε ελέγξτε τα διαπιστευτήρια, φαίνεται να είναι λάθος.",
+    "Please specify the port, it could not be auto-detected." : "Παρακαλούμε ορίστε την θύρα, δεν ήταν εφικτό να εντοπιστεί αυτόματα.",
+    "Base DN could not be auto-detected, please revise credentials, host and port." : "Αδυναμία ανίχνευσης βασικού DN, παρακαλούμε αναθεωρήστε τα διαπιστευτήρια, το διακομιστή και τη θύρα.",
+    "Could not detect Base DN, please enter it manually." : "Αδυναμία ανίχνευσης βασικού DN, παρακαλούμε να το εισάγετε χειροκίνητα.",
     "{nthServer}. Server" : "{nthServer}. Διακομιστής",
-    "No object found in the given Base DN. Please revise." : "Δεν βρέθηκε αντικείμενο στο δηλωθέν Base DN. Παρακαλώ αναθεωρήστε.",
-    "More than 1,000 directory entries available." : "Είναι διαθέσιμες περισσότερες από 1.000 εγγραφές καταλόγου.",
-    " entries available within the provided Base DN" : "διαθέσιμες καταχωρήσεις εντός του δηλωθέντος ",
-    "An error occurred. Please check the Base DN, as well as connection settings and credentials." : "Παρουσιάστηκε σφάλμα. Παρακαλούμε ελέγξτε το Base DN καθώς και τις ρυθμίσεις και τα διαπιστευτήρια σύνδεσης.",
+    "No object found in the given Base DN. Please revise." : "Δεν βρέθηκε αντικείμενο στο δηλωθέν βασικό DN. Παρακαλούμε αναθεωρήστε.",
+    "More than 1,000 directory entries available." : "Είναι διαθέσιμες περισσότερες από 1000 καταχωρήσεις καταλόγου.",
+    " entries available within the provided Base DN" : " διαθέσιμες καταχωρήσεις εντός του δηλωθέντος βασικού DN",
+    "An error occurred. Please check the Base DN, as well as connection settings and credentials." : "Παρουσιάστηκε σφάλμα. Παρακαλούμε ελέγξτε το βασικό DN καθώς και τις ρυθμίσεις και τα διαπιστευτήρια σύνδεσης.",
     "Do you really want to delete the current Server Configuration?" : "Θέλετε να διαγράψετε τις τρέχουσες ρυθμίσεις του διακομιστή;",
-    "Confirm Deletion" : "Επιβεβαίωση Διαγραφής",
+    "Confirm Deletion" : "Επιβεβαίωση διαγραφής",
     "Mappings cleared successfully!" : "Η εκκαθάριση αντιστοιχιών ήταν επιτυχής!",
     "Error while clearing the mappings." : "Σφάλμα κατά την εκκαθάριση των αντιστοιχιών.",
-    "Anonymous bind is not allowed. Please provide a User DN and Password." : "Η ανώνυμη δέσμευση δεν επιτρέπεται. Παρακαλούμε δώστε ένα DN χρήστη και Κωδικό Πρόσβασης.",
+    "Anonymous bind is not allowed. Please provide a User DN and Password." : "Η ανώνυμη δέσμευση δεν επιτρέπεται. Παρακαλούμε δώστε ένα DN χρήστη και συνθηματικό.",
     "LDAP Operations error. Anonymous bind might not be allowed." : "Σφάλμα ενεργειών LDAP. Η ανώνυμη δέσμευση μπορεί να μην επιτρέπεται.",
-    "Saving failed. Please make sure the database is in Operation. Reload before continuing." : "Αποτυχία αποθήκευσης. Παρακαλούμε βεβαιωθείτε ότι η βάση δεδομένων λειτουργεί. Επαναφορτώστε πριν συνεχίσετε.",
+    "Saving failed. Please make sure the database is in Operation. Reload before continuing." : "Αποτυχία αποθήκευσης. Παρακαλούμε βεβαιωθείτε ότι η βάση δεδομένων λειτουργεί. Φορτώστε ξανά πριν συνεχίσετε.",
     "Switching the mode will enable automatic LDAP queries. Depending on your LDAP size they may take a while. Do you still want to switch the mode?" : "Η αλλαγή της κατάστασης θα ενεργοποιήσει αυτόματα ερωτήματα LDAP. Ανάλογα με το μέγεθος του LDAP αυτό μπορεί να διαρκέσει αρκετά. Θέλετε ακόμη να αλλάξετε κατάσταση λειτουργίας;",
     "Mode switch" : "Αλλαγή κατάστασης",
     "Select attributes" : "Επιλογή χαρακτηριστικών",
-    "User not found. Please check your login attributes and username. Effective filter (to copy-and-paste for command line validation): <br/>" : "Δεν βρέθηκε ο χρήστης. Παρακαλούμε ελέγξτε ότι τις ιδιότητες σύνδεσης και το όνομα χρήστη. Ενεργό φίλτρο (για αντιγραφή και επικόλληση για επικύρωση σε γραμμή εντολών): ",
+    "User not found. Please check your login attributes and username. Effective filter (to copy-and-paste for command line validation): <br/>" : "Δεν βρέθηκε ο χρήστης. Παρακαλούμε ελέγξτε ότι τις ιδιότητες σύνδεσης και το όνομα χρήστη. Ενεργό φίλτρο (για αντιγραφή και επικόλληση για επικύρωση σε γραμμή εντολών): <br/>",
     "User found and settings verified." : "Ο χρήστης βρέθηκε και οι ρυθμίσεις επιβεβαιώθηκαν.",
-    "Settings verified, but more than one user was found. Only the first will be able to login. Consider a more narrow filter." : "Οι ρυθμίσεις επαληθεύτηκαν, αλλά βρέθηκαν περισσότεροι από ένας χρήστες. Μόνο ο πρώτος θα μπορέσει να συνδεθεί. Εξετάστε τη χρήση πιο επιλεκτικού φίλτρο.",
-    "An unspecified error occurred. Please check the settings and the log." : "Προέκυψε ένα απροσδιόριστο σφάλμα. Παρακαλούμε ελέγξτε τις ρυθμίσεςι και το αρχείο ακταγραφής.",
-    "The search filter is invalid, probably due to syntax issues like uneven number of opened and closed brackets. Please revise." : "Το φίλτρο αναζήτησης δεν είναι έγκυρο, πιθανώς λόγω συντακτικών προβλημάτων όπως μη ταίριασμα ανοίγματος και κλεισίματος αγκυλών. Παρακαλούμε διορθώστε.",
-    "A connection error to LDAP / AD occurred, please check host, port and credentials." : "Προέκυψε σφάλμα σύνδεσης στο LDAP / AD, παρακαλούμε ελέγξτε διακομιστή θύρα και διαπιστευτήρια.",
-    "The %uid placeholder is missing. It will be replaced with the login name when querying LDAP / AD." : "Η κράτηση θέσης %uid απουσιάζει. Θα αντικατασταθεί με το όνομα σύνδεσης κατά το ερώτημα ",
-    "Please provide a login name to test against" : "Παρακαλούμε δώστε ένα όνομα σύνδεσης για να γίνει δοκιμή",
+    "Settings verified, but more than one user was found. Only the first will be able to login. Consider a more narrow filter." : "Οι ρυθμίσεις επαληθεύτηκαν, αλλά βρέθηκαν περισσότεροι από ένας χρήστες. Μόνο ο πρώτος θα μπορέσει να συνδεθεί. Εξετάστε τη χρήση πιο επιλεκτικού φίλτρου.",
+    "An unspecified error occurred. Please check the settings and the log." : "Προέκυψε ένα απροσδιόριστο σφάλμα. Παρακαλούμε ελέγξτε τις ρυθμίσεις και το αρχείο καταγραφής.",
+    "The search filter is invalid, probably due to syntax issues like uneven number of opened and closed brackets. Please revise." : "Το φίλτρο αναζήτησης δεν είναι έγκυρο, πιθανώς λόγω συντακτικών θεμάτων όπως άνισο άνοιγμα και κλείσιμο αγκυλών. Παρακαλούμε διορθώστε.",
+    "A connection error to LDAP / AD occurred, please check host, port and credentials." : "Προέκυψε σφάλμα σύνδεσης στο LDAP / AD, παρακαλούμε ελέγξτε διακομιστή, θύρα και διαπιστευτήρια.",
+    "The %uid placeholder is missing. It will be replaced with the login name when querying LDAP / AD." : "Η κράτηση θέσης %uid απουσιάζει. Θα αντικατασταθεί με το όνομα σύνδεσης κατά το ερώτημα LDAP / AD.",
+    "Please provide a login name to test against" : "Παρακαλούμε δώστε ένα όνομα σύνδεσης για να γίνει δοκιμή ξανά",
     "The group box was disabled, because the LDAP / AD server does not support memberOf." : "Το πεδίο ομάδας απενεργοποιήθηκε επειδή ο διακομιστής LDAP / AD δεν υποστηρίζει το memberOf.",
     "Server" : "Διακομιστής",
     "Users" : "Χρήστες",
     "Login Attributes" : "Ιδιότητες Σύνδεσης",
     "Groups" : "Ομάδες",
-    "The configuration is invalid: anonymous bind is not allowed." : "Η διαμόρφωση δεν είναι έγκυρη: δεν επιτρέπεται ανώνυμη δέσμευση.",
+    "The configuration is invalid: anonymous bind is not allowed." : "Μη έγκυρη ρύθμιση: δεν επιτρέπεται ανώνυμη δέσμευση.",
     "The configuration is valid and the connection could be established!" : "Οι ρυθμίσεις είναι έγκυρες και η σύνδεση μπορεί να πραγματοποιηθεί!",
-    "The configuration is valid, but the Bind failed. Please check the server settings and credentials." : "Οι ρυθμίσεις είναι έγκυρες, αλλά απέτυχε η σύνδεση. Παρακαλώ ελέγξτε τις ρυθμίσεις του διακομιστή και τα διαπιστευτήρια.",
-    "The configuration is invalid. Please have a look at the logs for further details." : "Η διαμόρφωση είναι άκυρη. Παρακαλώ ελέγξτε τα αρχεία σφαλμάτων για περαιτέρω λεπτομέρειες.",
+    "The configuration is valid, but the Bind failed. Please check the server settings and credentials." : "Οι ρυθμίσεις είναι έγκυρες, αλλά απέτυχε η σύνδεση. Παρακαλούμε ελέγξτε τις ρυθμίσεις του διακομιστή και τα διαπιστευτήρια.",
+    "The configuration is invalid. Please have a look at the logs for further details." : "Μη έγκυρες ρυθμίσεις. Παρακαλούμε ελέγξτε τα αρχεία σφαλμάτων για περαιτέρω λεπτομέρειες.",
     "Failed to delete the server configuration" : "Αποτυχία διαγραφής ρυθμίσεων διακομιστή",
     "Failed to clear the mappings." : "Αποτυχία εκκαθάρισης των αντιστοιχιών.",
     "No data specified" : "Δεν προσδιορίστηκαν δεδομένα",
-    " Could not set configuration %s" : "Αδυναμία ρύθμισης %s",
+    " Could not set configuration %s" : " Αδυναμία ρύθμισης %s",
     "Action does not exist" : "Η ενέργεια δεν υπάρχει",
     "_%s group found_::_%s groups found_" : ["%s ομάδα βρέθηκε","%s ομάδες βρέθηκαν"],
     "_%s user found_::_%s users found_" : ["%s χρήστης βρέθηκε","%s χρήστες βρέθηκαν"],
-    "Could not detect user display name attribute. Please specify it yourself in advanced ldap settings." : "Δεν ήταν δυνατή η ανίχνευση της ιδιότητας του εμφανιζόμενου ονόματος χρήστη . Παρακαλείστε να το προσδιορίσετε στις προηγμένες ρυθμίσεις LDAP",
-    "Could not find the desired feature" : "Αδυναμία εύρεσης επιθυμητου χαρακτηριστικού",
-    "Invalid Host" : "Άκυρος εξυπηρετητής",
-    "Test Configuration" : "Δοκιμαστικες ρυθμισεις",
+    "Could not detect user display name attribute. Please specify it yourself in advanced ldap settings." : "Δεν ήταν δυνατή η ανίχνευση της ιδιότητας του εμφανιζόμενου ονόματος χρήστη. Παρακαλείστε να το προσδιορίσετε στις προηγμένες ρυθμίσεις ldap.",
+    "Could not find the desired feature" : "Αδυναμία εύρεσης επιθυμητού χαρακτηριστικού",
+    "Invalid Host" : "Μη έγκυρος διακομιστής",
+    "Test Configuration" : "Δοκιμαστικές ρυθμίσεις",
     "Help" : "Βοήθεια",
     "Groups meeting these criteria are available in %s:" : "Οι ομάδες που πληρούν τα κριτήρια είναι διαθέσιμες σε %s:",
-    "Only these object classes:" : "Μόνο οι κλάσεις αντικειμένων:",
-    "Only from these groups:" : "Μόνο από τις ομάδες:",
+    "Only these object classes:" : "Μόνο αυτές οι κλάσεις αντικειμένων:",
+    "Only from these groups:" : "Μόνο από αυτές τις ομάδες:",
     "Search groups" : "Αναζήτηση ομάδων",
     "Available groups" : "Διαθέσιμες ομάδες",
     "Selected groups" : "Επιλεγμένες ομάδες",
-    "Edit LDAP Query" : "Επεξεργασία ερωτήματος ",
+    "Edit LDAP Query" : "Επεξεργασία ερωτήματος LDAP",
     "LDAP Filter:" : "Φίλτρο LDAP:",
     "The filter specifies which LDAP groups shall have access to the %s instance." : "Το φίλτρο καθορίζει ποιες ομάδες LDAP θα έχουν πρόσβαση στην εγκατάσταση %s.",
     "Verify settings and count groups" : "Επιβεβαίωση ρυθμίσεων και καταμέτρηση ομάδων",
@@ -70,32 +70,32 @@ OC.L10N.register(
     "LDAP / AD Username:" : "Όνομα χρήστη LDAP / AD:",
     "Allows login against the LDAP / AD username, which is either uid or samaccountname and will be detected." : "Επιτρέπει σύνδεση με το όνομα χρήστη στο LDAP / AD, το οποίο είναι είτε uid ή samaccountname και θα ανιχνευθεί.",
     "LDAP / AD Email Address:" : "Διεύθυνση ηλ. ταχυδρομείου LDAP / AD:",
-    "Allows login against an email attribute. Mail and mailPrimaryAddress will be allowed." : "Επιτρέπει με χρήση μια ιδιότητας email. Τα mail και mailPrimaryAddress θα επιτραπούν.",
+    "Allows login against an email attribute. Mail and mailPrimaryAddress will be allowed." : "Επιτρέπει με χρήση μια ιδιότητας ηλ. ταχυδρομείου. Τα mail και mailPrimaryAddress θα επιτραπούν.",
     "Other Attributes:" : "Άλλες Ιδιότητες:",
-    "Defines the filter to apply, when login is attempted. %%uid replaces the username in the login action. Example: \"uid=%%uid\"" : "Ορίζει το φίλτρο που θα εφαρμοστεί, όταν επιχειριθεί σύνδεση. Το %%uid αντικαθιστά το όνομα χρήστη κατά τη σύνδεση. Παράδειγμα: \"uid=%%uid\"",
-    "Test Loginname" : "Δοκιμή ",
+    "Defines the filter to apply, when login is attempted. %%uid replaces the username in the login action. Example: \"uid=%%uid\"" : "Ορίζει το φίλτρο που θα εφαρμοστεί, όταν επιχειρηθεί σύνδεση. Το %%uid αντικαθιστά το όνομα χρήστη κατά τη σύνδεση. Παράδειγμα: \"uid=%%uid\"",
+    "Test Loginname" : "Δοκιμή ονόματος σύνδεσης",
     "Verify settings" : "Επιβεβαίωση ρυθμίσεων",
     "1. Server" : "1. Διακομιστής",
     "%s. Server:" : "%s. Διακομιστής:",
-    "Add a new and blank configuration" : "Προσθέτει μια νέα και κενή διαμόρφωση",
-    "Copy current configuration into new directory binding" : "Αντιγραφή της τρέχουσας διαμόρφωσης σε νέο κατάλογο",
-    "Delete the current configuration" : "Διαγραφή τρέχουσας διαμόρφωσης",
+    "Add a new and blank configuration" : "Προσθέτει μια νέα και κενή ρύθμιση",
+    "Copy current configuration into new directory binding" : "Αντιγραφή της τρέχουσας ρύθμισης σε νέο κατάλογο",
+    "Delete the current configuration" : "Διαγραφή τρέχουσας ρύθμισης",
     "Host" : "Διακομιστής",
     "You can omit the protocol, except you require SSL. Then start with ldaps://" : "Μπορείτε να παραλείψετε το πρωτόκολλο, εκτός αν απαιτείται SSL. Σε αυτή την περίπτωση ξεκινήστε με ldaps://",
     "Port" : "Θύρα",
     "Detect Port" : "Ανίχνευση Θύρας",
-    "User DN" : "User DN",
-    "The DN of the client user with which the bind shall be done, e.g. uid=agent,dc=example,dc=com. For anonymous access, leave DN and Password empty." : "Το DN του χρήστη πελάτη με το οποίο θα πρέπει να γίνει η σύνδεση, π.χ. uid=agent,dc=example,dc=com. Για χρήση χωρίς πιστοποίηση, αφήστε το DN και τον Κωδικό κενά.",
+    "User DN" : "DN χρήστη",
+    "The DN of the client user with which the bind shall be done, e.g. uid=agent,dc=example,dc=com. For anonymous access, leave DN and Password empty." : "Το DN του χρήστη πελάτη με το οποίο θα πρέπει να γίνει η σύνδεση, π.χ. uid=agent,dc=example,dc=com. Για ανώνυμη πρόσβαση, αφήστε κενά το DN και το συνθηματικό.",
     "Password" : "Συνθηματικό",
-    "For anonymous access, leave DN and Password empty." : "Για ανώνυμη πρόσβαση, αφήστε κενά τα πεδία DN και Pasword.",
-    "One Base DN per line" : "Ένα DN Βάσης ανά γραμμή ",
-    "You can specify Base DN for users and groups in the Advanced tab" : "Μπορείτε να καθορίσετε το Base DN για χρήστες και ομάδες από την καρτέλα Προηγμένες ρυθμίσεις",
-    "Detect Base DN" : "Ανίχνευση Base DN",
-    "Test Base DN" : "Δοκιμή ",
-    "Avoids automatic LDAP requests. Better for bigger setups, but requires some LDAP knowledge." : "Αποφυγή αυτόματων αιτημάτων LDAP. Προτιμότερο για μεγαλύτερες εγκαταστάσεις, αλλά απαιτεί κάποιες γνώσεις  LDAP.",
-    "Manually enter LDAP filters (recommended for large directories)" : "Εισάγετε με μη αυτόματο τρόπο φίλτρα LDAP (προτείνεται για μεγάλους καταλόγους)",
+    "For anonymous access, leave DN and Password empty." : "Για ανώνυμη πρόσβαση, αφήστε κενά τα πεδία DN και συνθηματικό.",
+    "One Base DN per line" : "Ένα βασικό DN ανά γραμμή",
+    "You can specify Base DN for users and groups in the Advanced tab" : "Μπορείτε να καθορίσετε το βασικό DN για χρήστες και ομάδες από την καρτέλα Προηγμένες ρυθμίσεις",
+    "Detect Base DN" : "Ανίχνευση βασικού DN",
+    "Test Base DN" : "Δοκιμή βασικού DN",
+    "Avoids automatic LDAP requests. Better for bigger setups, but requires some LDAP knowledge." : "Αποφυγή αυτόματων αιτημάτων LDAP. Προτιμότερο για μεγαλύτερες εγκαταστάσεις, αλλά απαιτεί κάποιες γνώσεις LDAP.",
+    "Manually enter LDAP filters (recommended for large directories)" : "Χειροκίνητη εισαγωγή φίλτρων LDAP (προτείνεται για μεγάλους καταλόγους)",
     "%s access is limited to users meeting these criteria:" : "%s η πρόσβαση είναι περιορισμένη σε χρήστες που πληρούν αυτά τα κριτήρια:",
-    "The most common object classes for users are organizationalPerson, person, user, and inetOrgPerson. If you are not sure which object class to select, please consult your directory admin." : "Οι πιο συνηθισμένες κλάσεις αντικειμένων για τους χρήστες είναι οι organizationalPerson, person, user, και inetOrgPerson. Εάν δεν είσαστε σίγουροι πιο κλάση να επιλέξετε, παρακαλώ συμβουλευτείτε τον διαχειριστή του καταλόγου σας.",
+    "The most common object classes for users are organizationalPerson, person, user, and inetOrgPerson. If you are not sure which object class to select, please consult your directory admin." : "Οι πιο συνηθισμένες κλάσεις αντικειμένων για τους χρήστες είναι οι organizationalPerson, person, user, και inetOrgPerson. Εάν δεν είσαστε σίγουροι πια κλάση να επιλέξετε, παρακαλούμε συμβουλευτείτε τον διαχειριστή του καταλόγου σας.",
     "The filter specifies which LDAP users shall have access to the %s instance." : "Το φίλτρο καθορίζει ποιοι χρήστες LDAP θα έχουν πρόσβαση στην εγκατάσταση %s.",
     "Verify settings and count users" : "Επαλήθευση ρυθμίσεων και καταμέτρηση χρηστών",
     "Saving" : "Αποθήκευση",
@@ -104,61 +104,63 @@ OC.L10N.register(
     "LDAP" : "LDAP",
     "Expert" : "Ειδικός",
     "Advanced" : "Για προχωρημένους",
-    "<b>Warning:</b> Apps user_ldap and user_webdavauth are incompatible. You may experience unexpected behavior. Please ask your system administrator to disable one of them." : "<b>Προσοχή:</b> Οι εφαρμογές user_ldap και user_webdavauth είναι ασύμβατες. Μπορεί να αντιμετωπίσετε απρόβλεπτη συμπεριφορά. Παρακαλώ ζητήστε από τον διαχειριστή συστήματος να απενεργοποιήσει μία από αυτές.",
-    "<b>Warning:</b> The PHP LDAP module is not installed, the backend will not work. Please ask your system administrator to install it." : "<b>Προσοχή:</b> Το άρθρωμα PHP LDAP δεν είναι εγκατεστημένο και το σύστημα υποστήριξης δεν θα δουλέψει. Παρακαλώ ζητήστε από τον διαχειριστή συστήματος να το εγκαταστήσει.",
+    "<b>Warning:</b> Apps user_ldap and user_webdavauth are incompatible. You may experience unexpected behavior. Please ask your system administrator to disable one of them." : "<b>Προσοχή:</b> Οι εφαρμογές user_ldap και user_webdavauth είναι ασύμβατες. Μπορεί να αντιμετωπίσετε απρόβλεπτη συμπεριφορά. Παρακαλούμε ζητήστε από τον διαχειριστή συστήματος να απενεργοποιήσει μία από αυτές.",
+    "<b>Warning:</b> The PHP LDAP module is not installed, the backend will not work. Please ask your system administrator to install it." : "<b>Προσοχή:</b> Το άρθρωμα PHP LDAP δεν είναι εγκατεστημένο και το σύστημα υποστήριξης δεν θα δουλέψει. Παρακαλούμε ζητήστε από τον διαχειριστή συστήματος να το εγκαταστήσει.",
     "Connection Settings" : "Ρυθμίσεις Σύνδεσης",
-    "Configuration Active" : "Ενεργοποιηση ρυθμισεων",
-    "When unchecked, this configuration will be skipped." : "Όταν δεν είναι επιλεγμένο, αυτή η ρύθμιση θα πρέπει να παραλειφθεί. ",
-    "Backup (Replica) Host" : "Δημιουργία αντιγράφων ασφαλείας (Replica) Host ",
-    "Give an optional backup host. It must be a replica of the main LDAP/AD server." : "Δώστε μια προαιρετική εφεδρική υποδοχή. Πρέπει να είναι ένα αντίγραφο του κύριου LDAP / AD διακομιστη.",
-    "Backup (Replica) Port" : "Δημιουργία αντιγράφων ασφαλείας (Replica) Υποδοχη",
-    "Disable Main Server" : "Απενεργοποιηση του κεντρικου διακομιστη",
-    "Only connect to the replica server." : "Σύνδεση μόνο με το διακομιστή-αντίγραφο.",
+    "Configuration Active" : "Οι ρυθμίσεις είναι ενεργές",
+    "When unchecked, this configuration will be skipped." : "Όταν δεν είναι επιλεγμένο, αυτή η ρύθμιση θα πρέπει να παραλειφθεί.",
+    "Backup (Replica) Host" : "Εφεδρικός (Replica) διακομιστή",
+    "Give an optional backup host. It must be a replica of the main LDAP/AD server." : "Δώστε προαιρετικά έναν εφεδρικό διακομιστή. Πρέπει να είναι ένα αντίγραφο του κύριου LDAP / AD διακομιστή.",
+    "Backup (Replica) Port" : "Εφεδρική (Replica) θύρα",
+    "Disable Main Server" : "Απενεργοποίηση του κεντρικού διακομιστή",
+    "Only connect to the replica server." : "Σύνδεση μόνο με τον εφεδρικό διακομιστή.",
     "Turn off SSL certificate validation." : "Απενεργοποίηση επικύρωσης πιστοποιητικού SSL.",
     "Not recommended, use it for testing only! If connection only works with this option, import the LDAP server's SSL certificate in your %s server." : "Δεν προτείνεται, χρησιμοποιείστε το μόνο για δοκιμές! Εάν η σύνδεση λειτουργεί μόνο με αυτή την επιλογή, εισάγετε το πιστοποιητικό SSL του διακομιστή LDAP στο %s διακομιστή σας.",
-    "Cache Time-To-Live" : "Cache Time-To-Live",
-    "in seconds. A change empties the cache." : "σε δευτερόλεπτα. Μια αλλαγή αδειάζει την μνήμη cache.",
+    "Cache Time-To-Live" : "Λανθάνουσα μνήμη Time-To-Live",
+    "in seconds. A change empties the cache." : "σε δευτερόλεπτα. Μια αλλαγή αδειάζει την λανθάνουσα μνήμη.",
+    "Network Timeout" : "Λήξη χρόνου δικτύου",
+    "timeout for all the ldap network operations, in seconds." : "λήξη ορίου για όλες τις λειτουργίες δικτύου ldap, σε δευτερόλεπτα.",
     "Directory Settings" : "Ρυθμίσεις Καταλόγου",
-    "User Display Name Field" : "Πεδίο Ονόματος Χρήστη",
-    "The LDAP attribute to use to generate the user's display name." : "Η ιδιότητα LDAP προς χρήση για δημιουργία του προβαλλόμενου ονόματος χρήστη.",
-    "2nd User Display Name Field" : "Πεδίο ονόματος 2ου ονόματος χρήστη",
+    "User Display Name Field" : "Πεδίο εμφανιζόμενου ονόματος χρήστη",
+    "The LDAP attribute to use to generate the user's display name." : "Η ιδιότητα LDAP προς χρήση για δημιουργία του εμφανιζόμενου ονόματος χρήστη.",
+    "2nd User Display Name Field" : "Πεδίο 2ου εμφανιζόμενου ονόματος χρήστη",
     "Optional. An LDAP attribute to be added to the display name in brackets. Results in e.g. »John Doe (john.doe@example.org)«." : "Προαιρετικά. Ένα χαρακτηριστικό LDAP θα προστεθεί στο εμφανιζόμενο όνομα σε παρενθέσεις. Τα αποτελέσματα θα είναι για παράδειγμα  »John Doe (john.doe@example.org)«.",
-    "Base User Tree" : "Base User Tree",
-    "One User Base DN per line" : "Ένα DN βάσης χρηστών ανά γραμμή",
-    "User Search Attributes" : "Χαρακτηριστικά αναζήτησης των χρηστών ",
-    "Optional; one attribute per line" : "Προαιρετικά? Ένα χαρακτηριστικό ανά γραμμή ",
+    "Base User Tree" : "Δέντρο βασικού χρήστη",
+    "One User Base DN per line" : "Ένα DN βάσης χρήστη ανά γραμμή",
+    "User Search Attributes" : "Χαρακτηριστικά αναζήτησης των χρηστών",
+    "Optional; one attribute per line" : "Προαιρετικά. Ένα χαρακτηριστικό ανά γραμμή",
     "Each attribute value is truncated to 191 characters" : "Κάθε τιμή χαρακτηριστικού περικόπτεται σε 191 χαρακτήρες",
-    "Group Display Name Field" : "Group Display Name Field",
-    "The LDAP attribute to use to generate the groups's display name." : "Η ιδιότητα LDAP προς χρήση για δημιουργία του προβαλλόμενου ονόματος ομάδας.",
-    "Base Group Tree" : "Base Group Tree",
-    "One Group Base DN per line" : "Μια ομαδικη Βάση DN ανά γραμμή",
-    "Group Search Attributes" : "Ομάδα Χαρακτηριστικων Αναζήτηση",
-    "Group-Member association" : "Group-Member association",
+    "Group Display Name Field" : "Πεδίο εμφανιζόμενου ονόματος ομάδας",
+    "The LDAP attribute to use to generate the groups's display name." : "Η ιδιότητα LDAP προς χρήση για δημιουργία του εμφανιζόμενου ονόματος ομάδας.",
+    "Base Group Tree" : "Δέντρο βασικής ομάδας",
+    "One Group Base DN per line" : "Ένα DN βάσης ομάδας ανά γραμμή",
+    "Group Search Attributes" : "Χαρακτηριστικά αναζήτησης των ομάδων",
+    "Group-Member association" : "Σχέση ομάδας-μέλους",
     "Dynamic Group Member URL" : "Δυναμικό URL Μέλους Ομάδας",
-    "The LDAP attribute that on group objects contains an LDAP search URL that determines what objects belong to the group. (An empty setting disables dynamic group membership functionality.)" : "Το χαρακτηριστικό του LDAP που σε αντικείμενα ομάδας περιέχει μια διεύθυνση URL αναζήτησης του LDAP,  καθορίζει ποια αντικείμενα ανήκουν στην ομάδα. (Μια κενή ρύθμιση απενεργοποιεί τη λειτουργικότητα δυναμικής συμμετοχής σε ομάδες.)",
+    "The LDAP attribute that on group objects contains an LDAP search URL that determines what objects belong to the group. (An empty setting disables dynamic group membership functionality.)" : "Το χαρακτηριστικό του LDAP που σε αντικείμενα ομάδας περιέχει μια διεύθυνση URL αναζήτησης του LDAP, καθορίζει ποια αντικείμενα ανήκουν στην ομάδα. (Μια κενή ρύθμιση απενεργοποιεί τη λειτουργικότητα δυναμικής συμμετοχής σε ομάδες.)",
     "Nested Groups" : "Φωλιασμένες ομάδες",
-    "When switched on, groups that contain groups are supported. (Only works if the group member attribute contains DNs.)" : "Όταν ενεργοποιηθεί, οι ομάδες που περιέχουν ομάδες υποστηρίζονται. (Λειτουργεί μόνο αν το χαρακτηριστικό μέλους ομάδες περιέχει Διακεκριμένα Ονόματα.)",
+    "When switched on, groups that contain groups are supported. (Only works if the group member attribute contains DNs.)" : "Όταν ενεργοποιηθεί, υποστηρίζονται οι ομάδες που περιέχουν ομάδες. (Λειτουργεί μόνο αν το χαρακτηριστικό μέλους ομάδες περιέχει DN.)",
     "Paging chunksize" : "Μέγεθος σελιδοποίησης",
     "Chunksize used for paged LDAP searches that may return bulky results like user or group enumeration. (Setting it 0 disables paged LDAP searches in those situations.)" : "Μέγεθος τμήματος που χρησιμοποιείται για την σελιδοποίηση αναζητήσεων LDAP που μπορεί να επιστρέψουν πολλά δεδομένα, όπως απαρίθμηση χρηστών ή ομάδων. (Η τιμή 0 απενεργοποιεί την σελιδοποίηση των αναζητήσεων LDAP σε αυτές τις περιπτώσεις.)",
     "Special Attributes" : "Ειδικά Χαρακτηριστικά ",
-    "Quota Field" : "Ποσοσταση πεδιου",
+    "Quota Field" : "Πεδίο χρήσης αποθηκευτικού χώρου",
     "Leave empty for user's default quota. Otherwise, specify an LDAP/AD attribute." : "Αφήστε κενό για να έχει ο χρήστης τον προεπιλεγμένο χώρο αποθήκευσης. Διαφορετικά, ορίστε το χαρακτηριστικό γνώρισμα του LDAP/AD.",
-    "Quota Default" : "Προκαθισμενο πεδιο",
-    "Override default quota for LDAP users who do not have a quota set in the Quota Field." : "Αντικατάσταση της προεπιλεγμένης ποσόστωσης για χρήστες LDAP που δεν έχουν καθορίσει τη ποσόστωση στο πεδίο Quota.",
-    "Email Field" : "Email τυπος",
-    "Set the user's email from their LDAP attribute. Leave it empty for default behaviour." : "Ορίστε την ηλεκτρονική διεύθυνση του χρήστη από το χαρακτηριστικό γνώρισμα του LDAP. Αφήστε το κενό για προεπιλεγμένη συμπεριφορά. ",
-    "User Home Folder Naming Rule" : "Χρήστης Προσωπικόςφάκελος Ονομασία Κανόνας ",
-    "Leave empty for user name (default). Otherwise, specify an LDAP/AD attribute." : "Αφήστε το κενό για το όνομα χρήστη (προεπιλογή). Διαφορετικά, συμπληρώστε μία ιδιότητα LDAP/AD.",
-    "Internal Username" : "Εσωτερικό Όνομα Χρήστη",
-    "By default the internal username will be created from the UUID attribute. It makes sure that the username is unique and characters do not need to be converted. The internal username has the restriction that only these characters are allowed: [ a-zA-Z0-9_.@- ].  Other characters are replaced with their ASCII correspondence or simply omitted. On collisions a number will be added/increased. The internal username is used to identify a user internally. It is also the default name for the user home folder. It is also a part of remote URLs, for instance for all *DAV services. With this setting, the default behavior can be overridden. To achieve a similar behavior as before ownCloud 5 enter the user display name attribute in the following field. Leave it empty for default behavior. Changes will have effect only on newly mapped (added) LDAP users." : "Εξ ορισμού, το εσωτερικό όνομα χρήστη θα δημιουργηθεί από το χαρακτηριστικό UUID. Αυτό βεβαιώνει ότι το όνομα χρήστη είναι μοναδικό και δεν χρειάζεται μετατροπή χαρακτήρων. Το εσωτερικό όνομα χρήστη έχει τον περιορισμό ότι μόνο αυτοί οι χαρακτήρες επιτρέπονται: [ a-zA-Z0-9_.@- ]. Οι άλλοι χαρακτήρες αντικαθίστανται με τους αντίστοιχους ASCII ή απλά παραλείπονται. Στις συγκρούσεις ένας αριθμός θα προστεθεί / αυξηθεί. Το εσωτερικό όνομα χρήστη χρησιμοποιείται για την αναγνώριση ενός χρήστη εσωτερικά. Είναι επίσης το προεπιλεγμένο όνομα για τον αρχικό φάκελο χρήστη. Αποτελεί επίσης μέρος των απομακρυσμένων διευθύνσεων URL, για παράδειγμα για όλες τις υπηρεσίες *DAV. Με αυτή τη ρύθμιση, η προεπιλεγμένη συμπεριφορά μπορεί να παρακαμφθεί. Για να επιτευχθεί μια παρόμοια συμπεριφορά όπως πριν το ownCloud 5 εισάγετε το χαρακτηριστικό του προβαλλόμενου ονόματος χρήστη στο παρακάτω πεδίο. Αφήστε το κενό για την προεπιλεγμένη λειτουργία. Οι αλλαγές θα έχουν ισχύ μόνο σε νεώτερους (προστιθέμενους) χρήστες LDAP.",
-    "Internal Username Attribute:" : "Ιδιότητα Εσωτερικού Ονόματος Χρήστη:",
+    "Quota Default" : "Προκαθορισμένη χρήσης αποθηκευτικού χώρου",
+    "Override default quota for LDAP users who do not have a quota set in the Quota Field." : "Αντικατάσταση της προεπιλεγμένου χώρου αποθήκευσης για χρήστες LDAP που δεν έχουν καθορίσει τον χώρο αποθήκευσης στο πεδίο Quota.",
+    "Email Field" : "Πεδίο ηλ. διεύθυνσης",
+    "Set the user's email from their LDAP attribute. Leave it empty for default behaviour." : "Ορίστε την ηλεκτρονική διεύθυνση του χρήστη από το χαρακτηριστικό γνώρισμα του LDAP. Αφήστε το κενό για προεπιλεγμένη συμπεριφορά.",
+    "User Home Folder Naming Rule" : "Κανόνας ονοματοθεσίας προσωπικού φακέλου χρήστη",
+    "Leave empty for user name (default). Otherwise, specify an LDAP/AD attribute." : "Αφήστε το κενό για το όνομα χρήστη (προεπιλογή). Διαφορετικά, συμπληρώστε ένα χαρακτηριστικό γνώρισμα του LDAP/AD.",
+    "Internal Username" : "Εσωτερικό όνομα χρήστη",
+    "By default the internal username will be created from the UUID attribute. It makes sure that the username is unique and characters do not need to be converted. The internal username has the restriction that only these characters are allowed: [ a-zA-Z0-9_.@- ].  Other characters are replaced with their ASCII correspondence or simply omitted. On collisions a number will be added/increased. The internal username is used to identify a user internally. It is also the default name for the user home folder. It is also a part of remote URLs, for instance for all *DAV services. With this setting, the default behavior can be overridden. To achieve a similar behavior as before ownCloud 5 enter the user display name attribute in the following field. Leave it empty for default behavior. Changes will have effect only on newly mapped (added) LDAP users." : "Εξ ορισμού, το εσωτερικό όνομα χρήστη θα δημιουργηθεί από το χαρακτηριστικό UUID. Αυτό βεβαιώνει ότι το όνομα χρήστη είναι μοναδικό και δεν χρειάζεται μετατροπή χαρακτήρων. Το εσωτερικό όνομα χρήστη έχει τον περιορισμό ότι μόνο αυτοί οι χαρακτήρες επιτρέπονται: [ a-zA-Z0-9_.@- ]. Οι άλλοι χαρακτήρες αντικαθίστανται με τους αντίστοιχους ASCII ή απλά παραλείπονται. Στις συγκρούσεις ένας αριθμός θα προστεθεί / αυξηθεί. Το εσωτερικό όνομα χρήστη χρησιμοποιείται για την αναγνώριση ενός χρήστη εσωτερικά. Είναι επίσης το προεπιλεγμένο όνομα για τον αρχικό φάκελο χρήστη. Αποτελεί επίσης μέρος των απομακρυσμένων διευθύνσεων URL, για παράδειγμα για όλες τις υπηρεσίες *DAV. Με αυτή τη ρύθμιση, η προεπιλεγμένη συμπεριφορά μπορεί να παρακαμφθεί. Για να επιτευχθεί μια παρόμοια συμπεριφορά όπως πριν το ownCloud 5 εισάγετε το χαρακτηριστικό του εμφανιζόμενου ονόματος χρήστη στο παρακάτω πεδίο. Αφήστε το κενό για την προεπιλεγμένη λειτουργία. Οι αλλαγές θα έχουν ισχύ μόνο σε νεώτερους (προστιθέμενους) χρήστες LDAP.",
+    "Internal Username Attribute:" : "Χαρακτηριστικό εσωτερικού ονόματος χρήστη:",
     "Override UUID detection" : "Παράκαμψη ανίχνευσης UUID",
     "By default, the UUID attribute is automatically detected. The UUID attribute is used to doubtlessly identify LDAP users and groups. Also, the internal username will be created based on the UUID, if not specified otherwise above. You can override the setting and pass an attribute of your choice. You must make sure that the attribute of your choice can be fetched for both users and groups and it is unique. Leave it empty for default behavior. Changes will have effect only on newly mapped (added) LDAP users and groups." : "Από προεπιλογή, το χαρακτηριστικό UUID εντοπίζεται αυτόματα. Το χαρακτηριστικό UUID χρησιμοποιείται για την αναγνώριση χωρίς αμφιβολία χρηστών και ομάδων LDAP. Επίσης, το εσωτερικό όνομα χρήστη θα δημιουργηθεί με βάση το UUID, εφόσον δεν ορίζεται διαφορετικά ανωτέρω. Μπορείτε να παρακάμψετε τη ρύθμιση και να ορίσετε ένα χαρακτηριστικό της επιλογής σας. Θα πρέπει να βεβαιωθείτε ότι το χαρακτηριστικό της επιλογής σας μπορεί να ληφθεί για τους χρήστες και τις ομάδες και ότι είναι μοναδικό. Αφήστε το κενό για την προεπιλεγμένη λειτουργία. Οι αλλαγές θα έχουν ισχύ μόνο σε πρόσφατα αντιστοιχισμένους (προστιθέμενους) χρήστες και ομάδες LDAP.",
-    "UUID Attribute for Users:" : "Χαρακτηριστικό UUID για Χρήστες:",
-    "UUID Attribute for Groups:" : "Χαρακτηριστικό UUID για Ομάδες:",
-    "Username-LDAP User Mapping" : "Αντιστοίχιση Χρηστών Όνομα Χρήστη-LDAP",
+    "UUID Attribute for Users:" : "Χαρακτηριστικό UUID για χρήστες:",
+    "UUID Attribute for Groups:" : "Χαρακτηριστικό UUID για ομάδες:",
+    "Username-LDAP User Mapping" : "Αντιστοίχιση χρήστη LDAP - όνομα χρήστη",
     "Usernames are used to store and assign (meta) data. In order to precisely identify and recognize users, each LDAP user will have an internal username. This requires a mapping from username to LDAP user. The created username is mapped to the UUID of the LDAP user. Additionally the DN is cached as well to reduce LDAP interaction, but it is not used for identification. If the DN changes, the changes will be found. The internal username is used all over. Clearing the mappings will have leftovers everywhere. Clearing the mappings is not configuration sensitive, it affects all LDAP configurations! Never clear the mappings in a production environment, only in a testing or experimental stage." : "Τα ονόματα χρηστών χρησιμοποιούνται για την αποθήκευση και τον προσδιορισμό των  (μετα) δεδομένων. Προκειμένου να προσδιοριστούν με ακρίβεια και να αναγνωριστούν οι χρήστες, κάθε χρήστης LDAP θα έχει ένα εσωτερικό όνομα. Αυτό απαιτεί μια αντιστοίχιση του ονόματος χρήστη με το χρήστη LDAP. Το όνομα χρήστη που δημιουργήθηκαν αντιστοιχίζεται στην UUID του χρήστη LDAP. Επιπροσθέτως, το DN αποθηκεύεται προσωρινά (cache) ώστε να μειωθεί η αλληλεπίδραση LDAP, αλλά δεν χρησιμοποιείται για την ταυτοποίηση. Αν το DN αλλάξει, οι αλλαγές θα εντοπιστούν. Το εσωτερικό όνομα χρήστη χρησιμοποιείται παντού. Η εκκαθάριση των αντιστοιχίσεων θα αφήσει υπολείμματα παντού. Η εκκαθάριση των αντιστοιχίσεων δεν επηρεάζεται από τη διαμόρφωση, επηρεάζει όλες τις διαμορφώσεις LDAP! Μην διαγράψετε ποτέ τις αντιστοιχίσεις σε ένα λειτουργικό περιβάλλον παρά μόνο σε δοκιμές ή σε πειραματικό στάδιο.",
-    "Clear Username-LDAP User Mapping" : "Διαγραφή αντιστοίχησης Ονόματος Χρήστη LDAP-Χρήστη",
-    "Clear Groupname-LDAP Group Mapping" : "Διαγραφή αντιστοίχησης Ονόματος Ομάδας-LDAP Ομάδας"
+    "Clear Username-LDAP User Mapping" : "Εκκαθάριση αντιστοίχησης χρήστη LDAP - όνομα χρήστη",
+    "Clear Groupname-LDAP Group Mapping" : "Εκκαθάριση αντιστοίχησης ονόματος ομάδας - ομάδας LDAP"
 },
 "nplurals=2; plural=(n != 1);");
