@@ -33,7 +33,7 @@ OC.L10N.register(
     "An unspecified error occurred. Please check the settings and the log." : "Produciuse un erro non agardado. Comprobe os axustes e o rexistro.",
     "The search filter is invalid, probably due to syntax issues like uneven number of opened and closed brackets. Please revise." : "O filtro de busca é incorrecto, probabelmente por mor de erros de sintaxe como un número impar de chaves de apertura/peche. Revíseo.",
     "A connection error to LDAP / AD occurred, please check host, port and credentials." : "Produciuse un erro de conexión no LDAP / AD, comprobe a máquina o porto e as credenciais.",
-    "The %uid placeholder is missing. It will be replaced with the login name when querying LDAP / AD." : "Non se atopou o marcador de posición %uid. Vai seren substituído co nome de acceso cando se consulta LDAP / AD.",
+    "The %uid placeholder is missing. It will be replaced with the login name when querying LDAP / AD." : "Non se atopou o marcador de posición %uid. Vai ser substituído co nome de acceso cando se consulta LDAP / AD.",
     "Please provide a login name to test against" : "Forneza o nome de acceso para facer a proba",
     "The group box was disabled, because the LDAP / AD server does not support memberOf." : "A caixa de grupo está desactivada, o servidor LDAP / AD non admite «memberOf».",
     "Server" : "Servidor",
@@ -51,7 +51,7 @@ OC.L10N.register(
     "Action does not exist" : "Non existe esta acción",
     "_%s group found_::_%s groups found_" : ["Atopouse %s grupo","Atopáronse %s grupos"],
     "_%s user found_::_%s users found_" : ["Atopouse %s usuario","Atopáronse %s usuarios"],
-    "Could not detect user display name attribute. Please specify it yourself in advanced ldap settings." : "Non foi posíbel detectar o atributo nome de usuario que mostrar. Especifíqueo vostede mesmo nos axustes avanzados de LDAP.",
+    "Could not detect user display name attribute. Please specify it yourself in advanced ldap settings." : "Non foi posíbel detectar o atributo nome de usuario que amosar. Especifíqueo vostede mesmo nos axustes avanzados de LDAP.",
     "Could not find the desired feature" : "Non foi posíbel atopar a función desexada",
     "Invalid Host" : "Máquina incorrecta",
     "Test Configuration" : "Probar a configuración",
@@ -104,14 +104,14 @@ OC.L10N.register(
     "LDAP" : "LDAP",
     "Expert" : "Experto",
     "Advanced" : "Avanzado",
-    "<b>Warning:</b> Apps user_ldap and user_webdavauth are incompatible. You may experience unexpected behavior. Please ask your system administrator to disable one of them." : "<b>Aviso:</b> Os aplicativos user_ldap e user_webdavauth son incompatíbeis. Pode acontecer un comportamento estraño. Consulte co administrador do sistema para desactivar unha delas.",
+    "<b>Warning:</b> Apps user_ldap and user_webdavauth are incompatible. You may experience unexpected behavior. Please ask your system administrator to disable one of them." : "<b>Aviso:</b> As aplicacións user_ldap e user_webdavauth son incompatíbeis. Pode acontecer un comportamento estraño. Consulte co administrador do sistema para desactivar unha delas.",
     "<b>Warning:</b> The PHP LDAP module is not installed, the backend will not work. Please ask your system administrator to install it." : "<b>Aviso:</b> O módulo PHP LDAP non está instalado, o servidor non funcionará. Consulte co administrador do sistema para instalalo.",
     "Connection Settings" : "Axustes da conexión",
     "Configuration Active" : "Configuración activa",
     "When unchecked, this configuration will be skipped." : "Se está sen marcar, omítese esta configuración.",
-    "Backup (Replica) Host" : "Servidor da copia de seguranza (Réplica)",
-    "Give an optional backup host. It must be a replica of the main LDAP/AD server." : "Indicar un servidor de copia de seguranza opcional. Debe ser unha réplica do servidor principal LDAP/AD.",
-    "Backup (Replica) Port" : "Porto da copia de seguranza (Réplica)",
+    "Backup (Replica) Host" : "Servidor da copia de seguridade (réplica)",
+    "Give an optional backup host. It must be a replica of the main LDAP/AD server." : "Indicar un servidor de copia de seguridade opcional. Debe ser unha réplica do servidor principal LDAP/AD.",
+    "Backup (Replica) Port" : "Porto da copia de seguridade (réplica)",
     "Disable Main Server" : "Desactivar o servidor principal",
     "Only connect to the replica server." : "Conectar só co servidor de réplica.",
     "Turn off SSL certificate validation." : "Desactiva a validación do certificado SSL.",
