@@ -91,7 +91,7 @@ class ExceptionOnLostConnection {
 	 *
 	 * @throws \Exception
 	 */
-	public function setUp() {
+	public function setUp(): void {
 		require_once __DIR__  . '/../../../../lib/base.php';
 		\OC_App::loadApps('user_ldap');
 
